add svg previews of images remove strange transitions on loading fix some color bugs

+++ sync.py
@@ -5,6 +5,7 @@ import urllib.request
 import urllib.error
 import json
 import os.path
+import subprocess
 from pprint import pprint
 
 def jget(url):
@@ -32,9 +33,6 @@ while uploads:
         if not 'cloudcasts' in rel:
             continue
         rel = rel['cloudcasts'][0]
-        if os.path.exists("content/mixes/%s.md" %rel['slug']):
-            print("content exists, skipping.")
-            continue
         if not 'user' in rel or \
             rel['user']["username"] not in VALID_USERS:
             continue
@@ -51,16 +49,30 @@ while uploads:
         data['date'] = rel['created_time']
         data['draft'] = False
         data['mixcloud'] = rel['slug']
+        data['img_png'] = os.path.join( "static", "images", "mixes", "%s.%s" %(rel['slug'], "png"))
+        data['img_svg'] = os.path.join("static", "images", "mixes", "%s.%s" %(rel['slug'], "svg"))
         
         pprint(data)
         output = template.format(
                     data=json.dumps(data, indent=2),
                     description=xd['description']
                     )
-        with open('content/mixes/%s.md' %rel['slug'], 'w') as fp:
-            fp.write(output)
-        print(output)
+        if os.path.exists("content/mixes/%s.md" %rel['slug']):
+            print("content exists, skipping.")
+        else:
+            with open('content/mixes/%s.md' %rel['slug'], 'w') as fp:
+                fp.write(output)
+        # print(output)
+        # download cover
+        if not os.path.exists(data['img_png']):
+            rq = urllib.request.urlopen(rel['pictures']['extra_large'])
+            with open(data['img_png'], "wb") as fp:
+                fp.write(rq.read())
+        if not os.path.exists(data['img_svg']):
+            subprocess.call(["node_modules/.bin/sqip", "-n", "4", "-b", "12",
+                             "-o", data['img_svg'], data['img_png']])
 
+        
     if not 'paging' in uploads or \
        not 'next' in uploads['paging']:
         break
